+++ CHANGES.md
@@ -1,9 +1,15 @@
 # Changes in this Release
 
+## v3.1.0
+
+- Add logo support
+- Remove default chart legend and merge it with the caption table
+- String changes
+- Possibility for rendering additional "Displacement max" column in the caption table
+
 ## v3.0.0
 
 - Use Node12 as it is supported on Windows7
-
 
 ## v2.0.0